requirements.txt
streamlit>=1.28.0,<1.35.0
faiss-cpu>=1.7.4,<1.8.0
sentence-transformers>=2.2.2,<2.8.0
google-generativeai>=0.3.0,<0.8.0
numpy>=1.24.0,<1.27.0
pandas>=2.0.0,<2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
PyMuPDF>=1.23.0,<1.25.0
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.10,<0.3.0
datasets>=2.14.0,<2.21.0
pyarrow>=10.0.0,<15.0.0
transformers>=4.25.0,<4.45.0
torch>=2.0.0,<2.5.0
